requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
typing_extensions==4.11.0
gunicorn
django-crispy-forms
crispy-bootstrap4